requirements-test.txt
django-test==0.4030
factory-boy==2.9.2
Faker==0.8.6
text-unidecode==1.0
pytest==3.2.3
pytest-forked==0.2
pytest-xdist==1.20.1
pytest-django==3.1.2
